ug_metal::utils

Trait EncoderParam

pub trait EncoderParam {
    // Required method
    fn set_param(encoder: &ComputeCommandEncoderRef, position: u64, data: Self);
}

Helper functions to create the various objects on the compute command encoder on a single line. Prevents getting wrong some arguments number and mixing length and size in bytes.

Required Methods

fn set_param(encoder: &ComputeCommandEncoderRef, position: u64, data: Self)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.